fix/ форма редактирования профиля под разные экраны
