+++ package.json
@@ -31,6 +31,7 @@
 		"react-dropdown-select": "^4.11.2",
 		"react-hook-form": "^7.50.1",
 		"react-redux": "^9.1.0",
+		"react-responsive": "^10.0.0",
 		"react-select": "^5.8.0",
 		"sass": "^1.70.0",
 		"scss": "^0.2.4",
