+++ src/entities/profile-edit-form/ui/profile-edit-form.module.scss
@@ -32,6 +32,7 @@
 	&_avatar {
 		background-color: $--bg-accent-inverse;
 		border-radius: 100px;
+		border: none;
 		height: 136px;
 		width: 136px;
 		overflow: hidden;
@@ -205,25 +206,25 @@
 	}
 }
 
-@media screen and (max-width: 799px) {
+@media screen and (max-width: 779px) {
 	.fields {
 		width: 328px;
 		padding: 16px;
 		&_photo {
 			position: relative;
 			display: flex;
-			justify-content: space-around;
+			flex-direction: column;
 			align-self: stretch;
-			margin-bottom: 16px;
 			width: 100%;
 			height: 171px;
+			margin-bottom: 0;
 			&_descr {
 				color: $--Foreground-subtitle;
 				font-family: $--open-sans-font;
 				font-size: 14px;
 				font-weight: 400;
-				margin-left: 150px;
 				margin-top: 63px;
+				margin-left: 0;
 			}
 		}
 		&_avatar {
@@ -235,6 +236,49 @@
 			position: absolute;
 			top: 0;
 			left: 0;
+		}
+		& label {
+			margin-top: 0;
+			flex: 0 0 auto;
+			align-self: stretch;
+			width: 100%;
+			div {
+				border-radius: 12px;
+				//	border: 1px solid $--border-default;
+				min-height: 52px;
+			}
+		}
+		&_contacts {
+			width: 100%;
+			display: grid;
+			grid-template-columns: repeat(3, 1fr);
+			grid-template-rows: 2 auto;
+			gap: 8px;
+			:nth-child(1) {
+				grid-column: 1 / 4;
+			}
+			:nth-child(2) {
+				grid-column: 1 / 4;
+				& div {
+					margin-top: 31px;
+				}
+			}
+			:nth-child(3) {
+				grid-row: 3;
+				grid-column: 1;
+			}
+			:nth-child(4) {
+				grid-row: 3;
+				grid-column: 2/4;
+				justify-self: flex-end;
+				border: none;
+				background: none;
+				color: $--Foreground-link-default;
+				font-family: $--open-sans-font;
+				font-size: 16px;
+				font-weight: 400;
+				cursor: pointer;
+			}
 		}
 	}
 }
