+++ src/entities/profile-edit-form/ui/profile-edit-form.tsx
@@ -1,419 +1,25 @@
 'use client';
 
-import React, { useState, useEffect } from 'react';
-import { MainButton, Input, Form } from '@/shared/ui';
-import { DatePickerRHF } from '@/shared/ui/date-picker-rhf/date-picker-rhf';
-import { useForm } from 'react-hook-form';
-import { PreviewProfile } from '@/shared/ui/preview-profile/preview-profile';
-import { ToggleCheckbox } from '@/shared/ui/toggle-checkbox/toggle-checkbox';
-import { ProfileLink } from '@/shared/ui/profile-link/profile-link';
-import Edit from '@/shared/assets/icons/edit-icon.svg';
-import Plus from '@/shared/assets/icons/plus-large.svg';
-import Photo from '@/shared/assets/icons/icon-photo.svg';
-import styles from './profile-edit-form.module.scss';
-import { PopUp } from '@/shared/ui';
-import { useMediaQuery } from '@/shared/hooks';
+import React, { useEffect } from 'react';
+import { useMediaQuery } from 'react-responsive';
+import { MobileView } from './pef-mobile';
+import { TabletView } from './pef-tablet';
+import { DesktopView } from './pef-desktop';
 
 export const ProfileEditForm = () => {
-	const [isPopup, setIsPopup] = useState(false);
-	const [preview, setPreview] = useState(false);
-	const isTablet = useMediaQuery('(max-width:1199px)');
-	const isMobile = useMediaQuery('(max-width:799px)');
-
-	const { control } = useForm();
-
-	const handlePreview = () => {
-		setPreview(true);
-	};
-	const saveAvatar = () => {
-		setIsPopup(false);
-	};
-	const handleSubmit = () => {
-		console.log('Submit');
-	};
+	const isDesktop = useMediaQuery({ query: '(min-width:1200px)' });
+	const isTablet = useMediaQuery({ minWidth: 780, maxWidth: 1199 });
+	const isMobile = useMediaQuery({ query: '(max-width:779px)' });
 
 	useEffect(() => {
 		window.scroll(0, 0);
 	}, []);
-
-	if (!isMobile && !isTablet) {
-		return (
-			<>
-				{preview ? (
-					<PreviewProfile />
-				) : (
-					<>
-						<ProfileLink title="Профиль" />
-						<Form onSubmit={handleSubmit} className={styles.fields}>
-							<div className={styles.fields_photo}>
-								<div className={styles.fields_avatar}>
-									<div className={styles.fields_text}>A</div>
-								</div>
-								<button
-									type="button"
-									className={styles.fields_edit}
-									onClick={() => setIsPopup(true)}>
-									<Edit />
-								</button>
-								<PopUp
-									visible={isPopup}
-									title="Изменить фото"
-									onClose={() => setIsPopup(false)}>
-									<Input
-										name="Foto"
-										type="file"
-										labelName="Изменить фото"></Input>
-									<MainButton
-										variant="primary"
-										width="regular"
-										onClick={saveAvatar}>
-										Сохранить
-									</MainButton>
-								</PopUp>
-								<div className={styles.fields_photo_descr}>
-									Загрузите файл в формате: JPEG, PNG, размером не более 10
-									Мбайт
-								</div>
-							</div>
-							<Input name="nick_name" labelName="Никнейм" description />
-							<Input
-								name="name"
-								labelName="Имя"
-								description={true}
-								descrText="Укажите свое настоящее имя и фамилию"
-							/>
-							<Input
-								name="about"
-								labelName="О себе"
-								description={true}
-								descrText="Не более 750 символов"
-							/>
-							<Input
-								name="portfolioLink"
-								labelName="Ссылка на портфолио"
-								description={true}
-								descrText="Добавьте ссылку на любую платформу, где размещено ваше портфолио"
-							/>
-							<div className={styles.fields_contacts}>
-								<Input
-									name="contacts"
-									labelName="Контакты для связи"
-									description={true}
-									descrText="Выберите ресурс"
-								/>
-								<Input name="source" labelName=" " />
-								<MainButton variant="secondary" width="regular" IconLeft={Plus}>
-									Добавить
-								</MainButton>
-								<button>Сбросить все</button>
-							</div>
-							<div className={styles.datePickerContainer}>
-								<p className={styles.datePickerTitle}>Дата рождения</p>
-								<DatePickerRHF control={control} name="birthDate" />
-							</div>
-							<div className={styles.fields_double}>
-								<Input name="country" labelName="Страна" description />
-								<Input name="city" labelName="Город" description />
-							</div>
-							<Input
-								name="speciality"
-								labelName="Специальность"
-								description={true}
-								descrText="Выберите не более 2 специальностей"
-							/>
-							<Input
-								name="skills"
-								labelName="Навыки"
-								description={true}
-								descrText="Выберите не более 15 навыков"
-							/>
-							<Input
-								name="qualLabel"
-								labelName="Уровень квалификации"
-								description
-							/>
-
-							<div className={styles.fields__checkbox}>
-								<p className={styles.fields__checkboxTitle}>
-									Готовность к участию в проектах
-								</p>
-								<div className={styles.fields__checkboxItem}>
-									<ToggleCheckbox
-										id="participation"
-										name="participation"
-										variant="defaultOf"
-										// checked={checked}
-										onChange={() => {}}
-									/>
-								</div>
-							</div>
-							<div className={styles.fields_buttonsContainer}>
-								<MainButton variant={'primary'} width={'regular'}>
-									Сохранить
-								</MainButton>
-								<button
-									type="button"
-									className={styles.button__preview}
-									onClick={handlePreview}>
-									{'Как видят мой профиль другие'}
-								</button>
-							</div>
-						</Form>
-					</>
-				)}
-			</>
-		);
-	} else if (isTablet) {
-		return (
-			<>
-				{preview ? (
-					<PreviewProfile />
-				) : (
-					<>
-						<ProfileLink title="Профиль" />
-						<Form onSubmit={handleSubmit} className={styles.fields}>
-							<div className={styles.fields_photo}>
-								<div className={styles.fields_avatar}>
-									<div className={styles.fields_text}>A</div>
-								</div>
-								<button
-									type="button"
-									className={styles.fields_edit}
-									onClick={() => setIsPopup(true)}>
-									<Edit />
-								</button>
-								<PopUp
-									visible={isPopup}
-									title="Изменить фото"
-									onClose={() => setIsPopup(false)}>
-									<Input
-										name="Foto"
-										type="file"
-										labelName="Изменить фото"></Input>
-									<MainButton
-										variant="primary"
-										width="regular"
-										onClick={saveAvatar}>
-										Сохранить
-									</MainButton>
-								</PopUp>
-								<div className={styles.fields_photo_descr}>
-									Загрузите файл в формате: JPEG, PNG, размером не более 10
-									Мбайт
-								</div>
-							</div>
-							<div className={styles.fields_double}>
-								<Input name="nick_name" labelName="Никнейм" description />
-								<Input
-									name="name"
-									labelName="Имя"
-									description={true}
-									descrText="Укажите свое настоящее имя и фамилию"
-								/>
-							</div>
-
-							<Input
-								name="about"
-								labelName="О себе"
-								description={true}
-								descrText="Не более 750 символов"
-							/>
-							<Input
-								name="portfolioLink"
-								labelName="Ссылка на портфолио"
-								description={true}
-								descrText="Добавьте ссылку на любую платформу, где размещено ваше портфолио"
-							/>
-							<div className={styles.fields_contacts}>
-								<Input
-									name="contacts"
-									labelName="Контакты для связи"
-									description={true}
-									descrText="Выберите ресурс"
-								/>
-								<Input name="source" labelName=" " />
-								<MainButton variant="secondary" width="regular" IconLeft={Plus}>
-									Добавить
-								</MainButton>
-								<button>Сбросить все</button>
-							</div>
-							<div className={styles.datePickerContainer}>
-								<p className={styles.datePickerTitle}>Дата рождения</p>
-								<DatePickerRHF control={control} name="birthDate" />
-							</div>
-
-							<Input name="country" labelName="Страна" description />
-							<Input name="city" labelName="Город" description />
-							<Input
-								name="speciality"
-								labelName="Специальность"
-								description={true}
-								descrText="Выберите не более 2 специальностей"
-							/>
-							<Input
-								name="skills"
-								labelName="Навыки"
-								description={true}
-								descrText="Выберите не более 15 навыков"
-							/>
-							<Input
-								name="qualLabel"
-								labelName="Уровень квалификации"
-								description
-							/>
-
-							<div className={styles.fields__checkbox}>
-								<p className={styles.fields__checkboxTitle}>
-									Готовность к участию в проектах
-								</p>
-								<div className={styles.fields__checkboxItem}>
-									<ToggleCheckbox
-										id="participation"
-										name="participation"
-										variant="defaultOf"
-										// checked={checked}
-										onChange={() => {}}
-									/>
-								</div>
-							</div>
-							<div className={styles.fields_buttonsContainer}>
-								<MainButton variant={'primary'} width={'regular'}>
-									Сохранить
-								</MainButton>
-								<button
-									type="button"
-									className={styles.button__preview}
-									onClick={handlePreview}>
-									{'Как видят мой профиль другие'}
-								</button>
-							</div>
-						</Form>
-					</>
-				)}
-			</>
-		);
-	} else if (isMobile) {
-		return (
-			<>
-				{preview ? (
-					<PreviewProfile />
-				) : (
-					<>
-						<ProfileLink title="Профиль" />
-						<Form onSubmit={handleSubmit} className={styles.fields}>
-							<button
-								className={styles.fields_photo}
-								type="button"
-								onClick={() => setIsPopup(true)}>
-								<Photo />
-							</button>
-							<PopUp
-								visible={isPopup}
-								title="Изменить фото"
-								onClose={() => setIsPopup(false)}>
-								<Input
-									name="Foto"
-									type="file"
-									labelName="Изменить фото"></Input>
-								<MainButton
-									variant="primary"
-									width="regular"
-									onClick={saveAvatar}>
-									Сохранить
-								</MainButton>
-							</PopUp>
-							<div className={styles.fields_photo_descr}>
-								Загрузите файл в формате: JPEG, PNG, размером не более 10 Мбайт
-							</div>
-							<div className={styles.fields_double}>
-								<Input name="nick_name" labelName="Никнейм" description />
-								<Input
-									name="name"
-									labelName="Имя"
-									description={true}
-									descrText="Укажите свое настоящее имя и фамилию"
-								/>
-							</div>
-
-							<Input
-								name="about"
-								labelName="О себе"
-								description={true}
-								descrText="Не более 750 символов"
-							/>
-							<Input
-								name="portfolioLink"
-								labelName="Ссылка на портфолио"
-								description={true}
-								descrText="Добавьте ссылку на любую платформу, где размещено ваше портфолио"
-							/>
-							<div className={styles.fields_contacts}>
-								<Input
-									name="contacts"
-									labelName="Контакты для связи"
-									description={true}
-									descrText="Выберите ресурс"
-								/>
-								<Input name="source" labelName=" " />
-								<MainButton variant="secondary" width="regular" IconLeft={Plus}>
-									Добавить
-								</MainButton>
-								<button>Сбросить все</button>
-							</div>
-							<div className={styles.datePickerContainer}>
-								<p className={styles.datePickerTitle}>Дата рождения</p>
-								<DatePickerRHF control={control} name="birthDate" />
-							</div>
-
-							<Input name="country" labelName="Страна" description />
-							<Input name="city" labelName="Город" description />
-							<Input
-								name="speciality"
-								labelName="Специальность"
-								description={true}
-								descrText="Выберите не более 2 специальностей"
-							/>
-							<Input
-								name="skills"
-								labelName="Навыки"
-								description={true}
-								descrText="Выберите не более 15 навыков"
-							/>
-							<Input
-								name="qualLabel"
-								labelName="Уровень квалификации"
-								description
-							/>
-
-							<div className={styles.fields__checkbox}>
-								<p className={styles.fields__checkboxTitle}>
-									Готовность к участию в проектах
-								</p>
-								<div className={styles.fields__checkboxItem}>
-									<ToggleCheckbox
-										id="participation"
-										name="participation"
-										variant="defaultOf"
-										// checked={checked}
-										onChange={() => {}}
-									/>
-								</div>
-							</div>
-							<div className={styles.fields_buttonsContainer}>
-								<MainButton variant={'primary'} width={'regular'}>
-									Сохранить
-								</MainButton>
-								<button
-									type="button"
-									className={styles.button__preview}
-									onClick={handlePreview}>
-									{'Как видят мой профиль другие'}
-								</button>
-							</div>
-						</Form>
-					</>
-				)}
-			</>
-		);
-	}
+	console.log(isDesktop, isTablet, isMobile);
+	return (
+		<>
+			{isDesktop && <DesktopView />}
+			{isTablet && <TabletView />}
+			{isMobile && <MobileView />}
+		</>
+	);
 };
