+++ src/widgets/footer/ui/footer.module.scss
@@ -84,7 +84,7 @@
 	}
 }
 
-@media screen and (max-width: 1200px) {
+@media screen and (max-width: 1199px) {
 	.footer__container {
 		margin: 0 auto;
 		max-width: 780px;
@@ -121,7 +121,7 @@
 	}
 }
 
-@media screen and (max-width: 780px) {
+@media screen and (max-width: 779px) {
 	.footer {
 		padding: 32px;
 		&__container {
